feat: custom date picker with calendar grid and full deadline formatting
- New DatePicker component: bottom-sheet calendar, Mon-first grid, 44px
  tap targets, terracotta selected date, weekends highlighted, past dates
  disabled, tomorrow as minimum date
- Deadline now shows as "15 июля 2026" in picker and "15 июля" in all
  order cards (feed, my-responses, orders list, respond form, detail page)

+++ app/feed/page.tsx
@@ -233,6 +233,6 @@ function getCategoryEmoji(cat: OrderCategory): string {
 }
 
 function formatDate(s: string) {
-  try { return new Date(s).toLocaleDateString('ru', { day: 'numeric', month: 'short' }); }
+  try { return new Date(s).toLocaleDateString('ru', { day: 'numeric', month: 'long' }); }
   catch { return s; }
 }
