+++ app/my-responses/page.tsx
@@ -17,7 +17,7 @@ interface Entry {
 }
 
 function formatDate(s: string) {
-  try { return new Date(s).toLocaleDateString('ru', { day: 'numeric', month: 'short' }); }
+  try { return new Date(s).toLocaleDateString('ru', { day: 'numeric', month: 'long' }); }
   catch { return s; }
 }
