+++ app/orders/[id]/page.tsx
@@ -212,6 +212,6 @@ export default function OrderDetailPage() {
 }
 
 function formatDate(s: string) {
-  try { return new Date(s).toLocaleDateString('ru', { day: 'numeric', month: 'short' }); }
+  try { return new Date(s).toLocaleDateString('ru', { day: 'numeric', month: 'long' }); }
   catch { return s; }
 }
