+++ app/orders/page.tsx
@@ -251,6 +251,6 @@ function OrderListCard({
 }
 
 function formatDate(s: string) {
-  try { return new Date(s).toLocaleDateString('ru', { day: 'numeric', month: 'short' }); }
+  try { return new Date(s).toLocaleDateString('ru', { day: 'numeric', month: 'long' }); }
   catch { return s; }
 }
